create a separate file position-tests.js for position test
